Factor out helper function into separate module.
